requirements.txt
matplotlib~=3.4.2
lmfit~=1.0.2
numpy~=1.20.3
PyQt5~=5.15.4
python-dateutil~=2.8.1
pandas~=1.2.4